# doc/a3s-openapi.yaml
openapi: 3.0.0
info:
  description: >-
    API Definition for the A3S. This service allows authentication, authorisation and accounting.
  version: 1.0.1
  title: A3S
paths:
...
  '/twoFactorAuth/users/{userId}':
    delete:
      tags:
        - TwoFactorAuth
      description: Removes two-factor authentication for a user.
      summary: Removes two-factor authentication for a user, as well as any registered devices and methods.
      operationId: removeTwoFactorAuthentication
      parameters:
        - name: userId
          in: path
          required: true
          description: The UUID of the user.
          schema:
            $ref: '#/components/schemas/UserId'
      responses:
        '400':
          description: Bad Request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Not authenticated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - You are not authorized to remove two-factor authentication.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: User not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        500:
          description: An unexpected error occurred.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  #endregion - TwoFactorAuth
  #region - Role
...
  '/securityContracts':
    get:
      tags:    
        - SecurityContract
      summary: Returns the entire security contract.
      description: Returns the entire security contract for the current state of A3S.
      operationId: getSecurityContract
      responses:
        '200':
          description: OK
          content:
            application/x-yaml:
              schema:
                $ref: '#/components/schemas/SecurityContract'
            application/json:
              schema:
                $ref: '#/components/schemas/SecurityContract'
        '204':
          description: No Content.
        '400':
          description: Bad Request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Not authenticated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - You are not authorized to access the security contract.
          content:
            application/json:
              schema:    
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Security contract not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: An unexpected error occurred.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      tags:
        - SecurityContract
      summary: Idempotently applies a security contract definition.
      description: Idempotently applies a security contract definition to the A3S instance.
      operationId: applySecurityContract
      requestBody:
        content:
          application/x-yaml:
            schema:
              $ref: '#/components/schemas/SecurityContract'
      responses:
        '200':
          description: OK.
          content:
            application/x-yaml:
              schema:
                $ref: '#/components/schemas/SecurityContract'
            application/json:
              schema:
                $ref: '#/components/schemas/SecurityContract'
        '204':
          description: No Content.
        '400':
          description: Bad Request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Not authenticated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Non-Processible Entity - The security contract was correctly structured, but there are business rule or constraint violations, preventing it from being applied.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - Not authorized to apply Security Contracts.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        500:
          description: An unexpected error occurred.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  # End Region - Security Contracts
    
  #region - AuthenticationModes    
...
components:
  schemas:
...
    UserId:
      type: string
      format: uuid
...
    Oauth2ClientSubmit:
      description: |
        Models a client definition required to create or update an Oauth2 or OpenID Connect client.
      type: object
      required:
        - clientId
        - name
        - allowedGrantTypes
        - allowedScopes
        - allowedOfflineAccess
        - redirectUris
        - allowedCorsOrigins
        - postLogoutRedirectUris
      properties:
        clientId:
          type: string
          description: The unique ID of the client.
        name:
          type: string
          description: Client display name. Used for logging and the consent screen.
        allowedGrantTypes:
          type: array
          items:
            type: string
          description: The Oauth2 grant types that the client will be permitted to use.
        redirectUris:
          description: Specifies the allowed URIs to return tokens or authorisation codes to.
          type: array
          items:
            type: string
            format: url
        allowedCorsOrigins:
          description: Sets the allowed CORS origins for JavaScript clients.
          type: array
          items:
            type: string
            format: url
        postLogoutRedirectUris:
          description: Specifies the allowed URIs to redirect to after logout.
          type: array
          items:
            type: string
            format: url
        allowedScopes:
          description: Specifies the scopes that the client is allowed to access. If empty, the client cannot access any scopes.
          type: array
          items:
            type: string
        clientSecrets:
          description: Client secrets for flows (grant types) that require one.
          type: array
          items:
            type: string
        hashedClientSecrets:
          description: An optional array containing already hashed client secrets for flows (grant types) that require one. If used, the clientSecrets will be ignored.
          type: array
          items:
            type: string
        allowedOfflineAccess:
          description: Defines whether offline access with refresh tokens is permitted for this client.
          type: boolean
        accessTokenLifetime:
          description: The amount of time that access tokens issued for this client will be valid (in seconds) for. Default is 3600.
          type: integer
        identityTokenLifetime:
          description: The amount of time that ID tokens issued for this client will be valid (in seconds) for. Default is 3600.
          type: integer
...
    SecurityContractPermission:
      description: |
          Models an application's permissions.
      type: object
      required:
        - name
        - description
      properties:
        name:
          type: string
          description: The name of the application permission.
        description:
          type: string
          description: A description of the application permission.
    SecurityContractFunction:
      description: |
          Models an application's functions. These contain permissions.
      type: object
      required:
        - name
        - description
        - permissions
      properties:
        name:
          type: string
          description: The name of the application function.
        description:
          type: string
          description: A description of the application function.
        permissions:
          type: array
          items:
            $ref: '#/components/schemas/SecurityContractPermission'
    SecurityContractApplication:
      description: |
        Models an application's functions and the corresponding permissions.
      type: object
      required:
        - fullname
        - applicationFunctions
      properties:
        fullname:
          type: string
          description: The name of the application function.
        applicationFunctions:
          type: array
          items:
            $ref: '#/components/schemas/SecurityContractFunction'
        dataPolicies:
          type: array
          items:
            $ref: '#/components/schemas/SecurityContractApplicationDataPolicy'
    SecurityContractApplicationDataPolicy:
      description: |
          Models an applications data policy with the context of security contracts.
      type: object
      required:
        - name
        - description
        - permissions
      properties:
        name:
          type: string
          description: The name of the application data policy.
        description:
          type: string
          description: A description of the application data policy.
        
    # End Region - Application Security Definition
    
    # Begin Region - Configuration Defaults
    SecurityContractDefaultConfigurationRole:
      description: |
          Models a role definition within the roles section of the default configuration.
      required:
        - name
        - description
      properties:
        name:
          type: string
          description: The name of the role.
        description:
          type: string
          description: A description of the role.
        functions:
          type: array
          description: An array of all the function names that are to be added to the role. The functions must already exist or be defined in other sections of the security contract.
          items:
            type: string
        roles:
          type: array
          description: An array of all the child roles that are to be added to the role. The roles must already exist or be defined in other sections of the security contract.
          items:
            type: string
    SecurityContractDefaultConfigurationUser:
      description: |
          Models a default user definition within a default configuration.
      required:
        - name
        - username
        - surname
        - email
      properties:
        uuid:
          type: string
          format: uuid
          description: The unique Id of the user.
        username:
          type: string
          description: The username of the user.
        name:
          type: string
          description: The name of the user.
        surname:
          type: string
          description: The surname of the user.
        email:
          type: string
          format: email
          description: The user's email.
        phoneNumber:
          type: string
          nullable: true
          description: The user's phone number.
        password:
          type: string
          description: The user's plain text password.
        hashedPassword:
          type: string
          description: An optional field containing the user's salted and hashed password. If used, the password field will be ignored.
        isDeleted:
          type: boolean
          description: True if this user is considered deleted.
        deletedTime:
          type: string
          format: date-time
          nullable: true
          description: Populated with user delete date and time.
        avatar:
          type: string
          nullable: true
          description: The user's avatar image in base64 format
        ldapAuthenticationMode:
          type: string
          nullable: true
          description: The user's linked LDAP Authentication Mode name, if applicable
        roles:
          type: array
          description: An array of all the role names that are to be added to the user. The user must already exist or be defined in other sections of the security contract.
          items:
            type: string
    SecurityContractDefaultConfigurationLdapAttributeLink:
      description: |
        An LdapAttributeLink - A LDAP attribute link definition
      type: object
      properties:
        userField:
          type: string
        ldapField:
          type: string
    SecurityContractDefaultConfigurationLdapAuthMode:
      description: |
        Model a default LDAP Authentication Mode
      required:
        - name
        - hostName
        - port
        - isLdaps
        - account
        - password
        - baseDn
      type: object
      properties:
        name:
          type: string
        hostName:
          type: string
        port:
          type: integer
          format: int32
        isLdaps:
          type: boolean
        account:
          type: string
        baseDn:
          type: string
        ldapAttributes:
          type: array
          description: A list of LDAP attribute to user field mappings
          items:
            $ref: '#/components/schemas/SecurityContractDefaultConfigurationLdapAttributeLink'
    SecurityContractDefaultConfigurationTeam:
      description: |
          Models a default team definition within a default configuration.
      required:
        - name
        - description
      properties:
        uuid:
          type: string
          format: uuid
          description: The unique Id of the team.
        name:
          type: string
          description: The name of the team.
        description:
          type: string
          description: A description of the team.
        users:
          type: array
          description: An array of all the user names that are to be added to the team. The user must already exist or be defined in other sections of the security contract.
          items:
            type: string
        teams:
          type: array
          description: An array of all the teams that are to be added to the team. The teams must already exist or be defined in other sections of the security contract.
          items:
            type: string
        dataPolicies:
          type: array
          description: An array of application data policy names that are to be applied to the team.
          items:
            type: string
    SecurityContractDefaultConfigurationApplication:
      required:
        - name
        - functions
      properties:
        name:
          type: string
          description: The name of the application that defaults are being configured for.
        functions:
          type: array
          items:
              $ref: '#/components/schemas/SecurityContractDefaultConfigurationFunction'
    SecurityContractDefaultConfigurationFunction:
      description: Models a section of the default configuration for the creation of default functions.
      type: object
      required:
        - name
        - description
        - permissions
      properties:
        name:
          type: string
          description: The name of the function.
        description:
          type: string
          description: A description of the function.
        permissions:
          type: array
          description: A list of permissions to add to the function.
          items:
            type: string 
    SecurityContractDefaultConfiguration:
      description: Models the default configuration section of a security contract.
      properties:
        name:
          type: string
          description: The name of the default configuration.
        applications:
          type: array
          items:
            $ref: '#/components/schemas/SecurityContractDefaultConfigurationApplication'
        roles:
          type: array
          items:
            $ref: '#/components/schemas/SecurityContractDefaultConfigurationRole'
        ldapAuthenticationModes:
          type: array
          items:
            $ref: '#/components/schemas/SecurityContractDefaultConfigurationLdapAuthMode'
        users:
          type: array
          items:
            $ref: '#/components/schemas/SecurityContractDefaultConfigurationUser'
        teams:
          type: array
          items:
            $ref: '#/components/schemas/SecurityContractDefaultConfigurationTeam'
    # End Region - Configuration Defaults
    
    # Begin Region - Security Contract
    SecurityContract:
      description: |
        Models a security contract definition.
      type: object
      properties:
        clients:
          type: array
          items:
            $ref: '#/components/schemas/Oauth2ClientSubmit'
        applications:
          type: array
          items:
             $ref: '#/components/schemas/SecurityContractApplication'
        defaultConfigurations:
          type: array
          items:
             $ref: '#/components/schemas/SecurityContractDefaultConfiguration'
    # End Region - Security Contract
    
    #region - Authentication Model
...
    ErrorResponse:
      description: Error response object to give more info with HTTP status code
      type: object
      required:
        - message
      properties:
        message:
          type: string
...